[refactor] DAO class table 단위로 분리 & DTO 분리 refactoring
